Move Wechat tests to Weixin
Move the Wechat-specific tests to the WeixinTests class to match the way the rest of the test project is organised.
